Update main server file
