+++ server/.env
@@ -1,1 +1,1 @@
-ATLAS_URI=mongodb+srv://Cluster25268:Password123@cluster25268.tsslojv.mongodb.net/
+ATLAS_URI=mongodb+srv://root:password12345678@cluster0.lfbgr2v.mongodb.net/?retryWrites=true&w=majority
