+++ server/package.json
@@ -2,9 +2,9 @@
     "name": "server",
     "version": "1.0.0",
     "description": "",
-    "main": "index.js",
+    "main": "server.js",
     "scripts": {
-        "dev": "nodemon ./src/index.js"
+        "dev": "nodemon server.js"
     },
     "keywords": [],
     "author": "",
